move field data to separate file
